Galaxy source v.0.3.1

+++ src/Global.hpp
@@ -2,7 +2,7 @@
 class Global
 {
 public:
-	int themeID = 0;
+	int8_t themeID = 0;
 	bool wantInput = false;
 	static Global& Get() { static Global g; return g; }
 };
